requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
uvicorn[standard]>=0.30
requests>=2.31
psycopg2-binary>=2.9
django-redis>=6.0
tusk-drift-python-sdk[django]>=0.1.19
